型定義ファイル. Styled-Components 導入.
