+++ src/components/layout/index.ts
@@ -1,3 +1,2 @@
 export { default as Content } from './Content';
-export { default as Header } from './Header';
-export { default as Sidebar } from './Sidebar';
+export { default as Layout } from './Layout';
