+++ src/model/index.ts
@@ -1,0 +1,3 @@
+export * from './project.model';
+export * from './task.model';
+export * from './user.model';
