+++ src/model/project.model.ts
@@ -1,0 +1,6 @@
+export type Project = {
+  id: string,
+  docId: string,
+  name: string,
+  userId: string,
+};
