+++ src/model/task.model.ts
@@ -1,0 +1,8 @@
+export type Task = {
+  id: string,
+  docId: string,
+  name: string;
+  projectId: string;
+  date: string;
+  archived: boolean;
+};
