+++ src/model/user.model.ts
@@ -1,0 +1,7 @@
+export type User = {
+  docId: string,
+  firstName: string;
+  lastName: string;
+  twitter: string;
+  userId: string;
+};
