+++ src/routes/Routes.tsx
@@ -1,27 +1,34 @@
+import {Layout} from '@/components/layout';
 import {ProjectView, TodayView} from '@/views';
 import {createBrowserRouter, Navigate, RouteObject} from 'react-router-dom';
 
 // ルートの設定.
 const routes: RouteObject[] = [
   {
-    // Base route.
     path: '/',
-    element: (
-      <Navigate
-        to="/today"
-        replace={true}
-      />
-    ),
-  },
-  // Today.
-  {
-    path: 'today',
-    element: <TodayView />,
-  },
-  // By project.
-  {
-    path: 'project/:projectId',
-    element: <ProjectView />,
+    element: <Layout />,
+    children: [
+      // Base route.
+      {
+        index: true,
+        element: (
+          <Navigate
+            to="/today"
+            replace={true}
+          />
+        ),
+      },
+      // Today.
+      {
+        path: 'today',
+        element: <TodayView />,
+      },
+      // By project.
+      {
+        path: 'project/:projectId',
+        element: <ProjectView />,
+      },
+    ],
   },
 ];
