--- src/types/componentProps.ts
@@ -1,9 +1,0 @@
-import { Project, Task } from './models';
-
-export interface ProjectListProps {
-  projects: Project[];
-}
-
-export interface TaskListProps {
-  tasks: Task[];
-}
