+++ src/views/ProjectView.tsx
@@ -1,26 +1,9 @@
-import { Content, Header, Sidebar } from '@/components/layout';
-import { ProjectsProvider, TaskFilterProvider } from '@/components/provider';
-import {useState} from 'react';
+import {Content} from '@/components/layout';
 
 export default function ProjectView(): React.ReactNode {
-  const [darkMode, setDarkMode] = useState<boolean>(false);
-
   return (
     <>
-      <TaskFilterProvider>
-        <ProjectsProvider>
-          <main className={darkMode ? 'darkmode' : ''}>
-            <Header
-              darkMode={darkMode}
-              setDarkMode={setDarkMode}
-            />
-            <section className="content">
-              <Sidebar />
-              <Content />
-            </section>
-          </main>
-        </ProjectsProvider>
-      </TaskFilterProvider>
+      <Content />
     </>
   );
 }
