+++ src/views/TodayView.tsx
@@ -1,26 +1,9 @@
-import {useState} from 'react';
-import { ProjectsProvider, TaskFilterProvider } from '@/components/provider';
-import { Content, Header, Sidebar } from '@/components/layout';
+import {Content} from '@/components/layout';
 
 export default function TodayView(): React.ReactNode {
-  const [darkMode, setDarkMode] = useState<boolean>(false);
-
   return (
     <>
-      <TaskFilterProvider>
-        <ProjectsProvider>
-          <main className={darkMode ? 'darkmode' : ''}>
-            <Header
-              darkMode={darkMode}
-              setDarkMode={setDarkMode}
-            />
-            <section className="content">
-              <Sidebar />
-              <Content />
-            </section>
-          </main>
-        </ProjectsProvider>
-      </TaskFilterProvider>
+      <Content />
     </>
   );
 }
